Prime.pm:  Don't check overlarge divisors
Since we're checking all primes in ascending order, a composite number's
least factor will be reached first, and there must be a greater or equal
factor, which multiplied to the first yields the number in question.  If
the square of a prime exceeds the number, then it isn't the least factor
(because the product of it and a greater or equal prime must also exceed
the number) and no greater prime can be the least factor either.  And we
only reach a prime by ruling out the lesser primes as factors, so we can
conclude that no primes are factors and the number itself is prime.

This yields dramatic performance improvements.